add string user for model Task, added django-crispy and update view funstion

+++ tasks/views.py
@@ -1,13 +1,15 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponseForbidden
 from django.template.loader import render_to_string
+from django.contrib.auth.decorators import login_required
 
 from tasks.forms import TaskForm
 from tasks.models import Task
 
 
+@login_required
 def get_task(request, task_id):
-    task = get_object_or_404(Task, id=task_id)
+    task = get_object_or_404(Task, id=task_id, user=request.user)
     if not task:
         raise ValueError("Данной задачи не найдено")
     context = {
@@ -17,14 +19,15 @@ def get_task(request, task_id):
     return render(request, 'tasks/task_detail.html', context=context)
 
 
+@login_required
 def get_tasks(request):
     status = request.GET.get('status')
     if status == 'completed':
-        tasks = Task.objects.filter(is_completed=True)
+        tasks = Task.objects.filter(is_completed=True, user=request.user)
     elif status == 'active':
-        tasks = Task.objects.filter(is_completed=False)
+        tasks = Task.objects.filter(is_completed=False, user=request.user)
     else:
-        tasks = Task.objects.all()
+        tasks = Task.objects.filter(user=request.user)
     
     if request.headers.get('x-requested-with') == 'XMLHttpRequest':
         html = render_to_string('tasks/includes/task_list_items.html', {'tasks': tasks})
@@ -36,11 +39,14 @@ def get_tasks(request):
     return render(request, 'tasks/task_list.html', context=context)
 
 
+@login_required
 def create_task(request):
     if request.method == 'POST':
         form = TaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            task = form.save(commit=False)
+            task.user = request.user
+            task.save()
             return redirect('tasks')
     else:
         form = TaskForm()
@@ -51,8 +57,11 @@ def create_task(request):
     return render(request, 'tasks/task_form.html', context=context)
 
 
+@login_required
 def update_task(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+    if task.user != request.user:
+        return HttpResponseForbidden('Вы не можете редактировать чужую задачу')
     if request.method == 'POST':
         form = TaskForm(request.POST, instance=task)
         if form.is_valid():
@@ -68,9 +77,12 @@ def update_task(request, task_id):
     return render(request, 'tasks/task_form.html', context=context)
 
 
+@login_required
 def toggle_task_status(request, task_id):
     if request.method == 'POST' and request.headers.get('x-requested-with') == 'XMLHttpRequest':
         task = get_object_or_404(Task, id=task_id)
+        if task.user != request.user:
+            return HttpResponseForbidden('Вы не можете редактировать чужую задачу')
         task.is_completed = not task.is_completed
         task.save()
         return JsonResponse({
@@ -81,8 +93,11 @@ def toggle_task_status(request, task_id):
     return JsonResponse({'status': 'error'}, status=400)
 
 
+@login_required
 def delete_task(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+    if task.user != request.user:
+        return HttpResponseForbidden('Вы не можете удалить чужую задачу')
     if request.method == 'POST':
         task.delete()
         return redirect('tasks')
